+++ src/clients/entitlements/storage/in-memory/in-memory.cache-key.utils.ts
@@ -1,0 +1,9 @@
+import { FeatureKey } from '../../types';
+
+export const ENTITLEMENTS_MAP_KEY = 'entitlements';
+export const PERMISSIONS_MAP_KEY = 'permissions';
+export const SRC_BUNDLES_KEY = 'src_bundles';
+
+export function getFeatureEntitlementKey(featKey: FeatureKey, tenantId: string, userId = ''): string {
+  return `${tenantId}:${userId}:${featKey}`;
+}