+++ src/clients/index.ts
@@ -2,3 +2,4 @@ export * from './audits';
 export * from './http';
 export * from './identity';
 export * from './hosted-login';
+export * from './entitlements';