+++ docs/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [5.1.0-alpha.3](https://github.com/frontegg/nodejs-sdk/compare/5.1.0-alpha.2...5.1.0-alpha.3) (2023-07-25)
+
+
+### Bug Fixes
+
+* **entitlements:** rename reason to justification ([a437b84](https://github.com/frontegg/nodejs-sdk/commit/a437b84213ebffbe17590f898cb0d6908869121a))
+
+# [5.1.0-alpha.2](https://github.com/frontegg/nodejs-sdk/compare/5.1.0-alpha.1...5.1.0-alpha.2) (2023-07-21)
+
+
+### Bug Fixes
+
+* **entitlements:** fixed the way of fetching user ID ([6df03ac](https://github.com/frontegg/nodejs-sdk/commit/6df03ac156434b4eab1e76126a8766f36360fd03))
+
+# [5.1.0-alpha.1](https://github.com/frontegg/nodejs-sdk/compare/5.0.0...5.1.0-alpha.1) (2023-07-17)
+
+
+### Features
+
+* **entitlements:** add support for entitlements ([ce857c9](https://github.com/frontegg/nodejs-sdk/commit/ce857c96c1359e5e6aa39c6a7986b1704ca44b50))
+
 # [5.0.0](https://github.com/frontegg/nodejs-sdk/compare/4.2.2...5.0.0) (2023-06-21)